// apps/frontend/lib/api/generated/models/Profile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WorkExperience } from "./WorkExperience";
export type Profile = {
	id?: string;
	summary?: string;
	workExperiences?: Array<WorkExperience>;
	skills?: Array<string>;
};